refactor: orgUserApi handler
