+++ packages/nocodb/src/lib/services/index.ts
@@ -20,3 +20,4 @@ export * as metaDiffService from './metaDiffService';
 export * as mapViewService from './mapViewService';
 export * as modelVisibilityService from './modelVisibilityService';
 export * as sharedBaseService from './sharedBaseService';
+export * as orgUserService from './orgUserService';
